TimeStamp.now() uses System.nanoSec(). The MockDataSource uses that.

+++ PVManager/src/main/java/gov/bnl/pvmanager/TimeDuration.java
@@ -39,7 +39,7 @@ public class TimeDuration {
      * @param nanoSec nanoseconds of the duration
      * @return a new duration
      */
-    public static TimeDuration nanos(int nanoSec) {
+    public static TimeDuration nanos(long nanoSec) {
         return new TimeDuration(nanoSec);
     }
